drop UserLessonRepository
