+++ app/Http/Controllers/Api/LearningController.php
@@ -28,7 +28,7 @@ class LearningController extends Controller
         UserLessonRepositoryInterface $userLessonRepository,
         Lesson $lesson
     ): JsonResponse {
-        $userLesson = $userLessonRepository->fetchUserLesson($this->user(), $lesson->id);
+        $userLesson = $userLessonRepository->fetchUserLesson($lesson->id);
         $userExercise = $learningService->fetchRandomExerciseOfLesson($userLesson, $this->user(), $request->previous_exercise_id);
         return $this->response($userExercise);
     }
